% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_JS_dist.R
\name{calc_JS_dist}
\alias{calc_JS_dist}
\title{Evaluate JS distance between two fitted (Dirichlet) mixture models}
\usage{
calc_JS_dist(mod1, mod2, nsim = 100, nperm = 100,
  show_progress_perm = FALSE)
}
\arguments{
\item{mod1, mod2}{cyl_mix objects (outputs of fit_cyl_mix) whose distance is of interest.}
}
\description{
Evaluate JS distance between two fitted (Dirichlet) mixture models
}
